Correct string to int conversion

+++ src/dale.h
@@ -83,5 +83,5 @@ void *hostdopen(const char *path);
 char *hostdread(void *dir);
 void hostdclose(void *dir);
 bool hostisdir(const char *path);
-void hostexec(char **cmds, char **msgs, size_t len, int jobs);
+void hostexec(char **cmds, char **msgs, size_t len, size_t jobs);
 char *hostexecout(const char *cmd);
